# server/docs/swagger.yaml
definitions:
...
  datatypes.JSONMap:
    additionalProperties: true
    type: object
...
  example.ExaFileUploadAndDownload:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      UpdatedAt:
        description: 更新时间
        type: string
      id:
        description: 主键ID
        type: integer
      key:
        description: 编号
        type: string
      name:
        description: 文件名
        type: string
      tag:
        description: 文件标签
        type: string
      url:
        description: 文件地址
        type: string
    type: object
...
  request.CIRun:
    properties:
      api_message_id:
        description: 消息ID
        type: integer
      callback_url:
        description: 回调地址
        type: string
      env:
        description: 环境ID
        type: integer
      key:
        description: 运行Key
        type: string
      other:
        allOf:
        - $ref: '#/definitions/datatypes.JSONMap'
        description: 其他参数
      project:
        description: 项目ID
        type: integer
      report:
        description: 报告ID
        type: integer
      secret:
        description: 项目Secret
        type: string
      tag:
        description: 标签ID
        type: integer
      task:
        description: 任务ID
        type: integer
      uuid:
        description: 项目UUID
        type: string
    type: object
...
  request.PageInfo:
    properties:
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
...
  response.ExaFileResponse:
    properties:
      file:
        $ref: '#/definitions/example.ExaFileUploadAndDownload'
    type: object
  response.FilePathResponse:
    properties:
      filePath:
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /ci/runTag:
    get:
      consumes:
      - application/json
      parameters:
      - description: 消息ID
        in: query
        name: api_message_id
        type: integer
      - description: 回调地址
        in: query
        name: callback_url
        type: string
      - description: 环境ID
        in: query
        name: env
        type: integer
      - description: 运行Key
        in: query
        name: key
        type: string
      - description: 项目ID
        in: query
        name: project
        type: integer
      - description: 报告ID
        in: query
        name: report
        type: integer
      - description: 项目Secret
        in: query
        name: secret
        type: string
      - description: 标签ID
        in: query
        name: tag
        type: integer
      - description: 任务ID
        in: query
        name: task
        type: integer
      - description: 项目UUID
        in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"data":{"success": true},"msg":"查询成功"}'
          schema:
            type: string
      summary: 获取CI运行的测试报告
    post:
      consumes:
      - application/json
      parameters:
      - description: tag, env, uuid, secret
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CIRun'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"data":{"key": "f2dc1396-1d3e-4f12-8b9a-f9d35e88cd7e","report":
            12},"msg":"运行成功"}'
          schema:
            type: string
      summary: CI运行标签
...
  /fileUploadAndDownload/breakpointContinue:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: an example for breakpoint resume, 断点续传示例
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 断点续传到服务器
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 断点续传到服务器
      tags:
      - ExaFileUploadAndDownload
  /fileUploadAndDownload/deleteFile:
    post:
      parameters:
      - description: 传入文件里面id即可
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/example.ExaFileUploadAndDownload'
      produces:
      - application/json
      responses:
        "200":
          description: 删除文件
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除文件
  /fileUploadAndDownload/findFile:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 上传文件完成
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 创建文件,返回包括文件路径
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.FilePathResponse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建文件
      tags:
      - ExaFileUploadAndDownload
  /fileUploadAndDownload/getFileList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 分页文件列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页文件列表
  /fileUploadAndDownload/removeChunk:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 删除缓存切片
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 删除切片
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除切片
      tags:
      - ExaFileUploadAndDownload
  /fileUploadAndDownload/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 上传文件示例
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传文件示例,返回包括文件详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ExaFileResponse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 上传文件示例
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"